Restructured API Interface
